+++ src/messages/layouts/MessageLayout.hpp
@@ -116,7 +116,7 @@ private:
     void updateBuffer(QPixmap *buffer, const MessagePaintContext &ctx);
 
     // Create new buffer if required, returning the buffer
-    QPixmap *ensureBuffer(QPainter &painter, int width);
+    QPixmap *ensureBuffer(QPainter &painter, int width, bool clear);
 
     // variables
     MessagePtr message_;